pyproject.toml
[project]
name = "hpc-sync"
version = "0.1.0"
description = "Project-local sync utility for mirroring folders to HPC systems"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1",
  "pyyaml>=6.0",
]

[project.scripts]
hpc-sync = "hpc_sync.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
